refactor: move calendar components to calendarView folder
