+++ src/components/Main.tsx
@@ -6,9 +6,11 @@
  * @contact intra: @xifeng
  */
 
-import BasicGrids from '@/components/BasicGrids';
-import BookingsLayer from '@/components/BookingsLayer';
-import CalendarHeader from '@/components/CalendarHeader';
+import { useRef } from 'react';
+
+import BasicGrids from '@/components/calendarView/BasicGrids';
+import BookingsLayer from '@/components/calendarView/BookingsLayer';
+import CalendarHeader from '@/components/calendarView/CalendarHeader';
 import RoomMap from '@/components/RoomMap';
 import { CELL_HEIGHT_PX, CELL_WIDTH_PX, OPEN_HOURS_IDX, TIME_SLOT_INTERVAL } from '@/config';
 import { styleGenerator } from '@/lib/tools';
@@ -16,30 +18,34 @@ import { styleGenerator } from '@/lib/tools';
 const slotsInAHour = 60 / TIME_SLOT_INTERVAL;
 const rows = (OPEN_HOURS_IDX[1] - OPEN_HOURS_IDX[0]) / slotsInAHour;
 
-const Main = () => (
-  <div
-    data-role='main-wrapper'
-    className='flex w-full items-start justify-start overflow-x-scroll lg:items-center lg:justify-center'
-  >
-    <div data-role='main' className='mx-4 mt-5 mb-12 h-fit w-fit'>
-      <RoomMap />
+const Main = () => {
+  const containerRef = useRef(null);
+  return (
+    <div
+      ref={containerRef}
+      data-role='main-wrapper'
+      className='flex w-full items-start justify-start overflow-x-scroll lg:items-center lg:justify-center'
+    >
+      <div data-role='main' className='mx-4 mt-5 mb-12 h-fit w-fit'>
+        <RoomMap />
 
-      {/* Calendar */}
-      <div data-role='calendar'>
-        <CalendarHeader />
+        {/* Calendar */}
+        <div data-role='calendar'>
+          <CalendarHeader />
 
-        {/* Calendar data */}
-        <div
-          data-role='calendar-data-container'
-          className='relative'
-          style={styleGenerator(CELL_WIDTH_PX * 8, CELL_HEIGHT_PX * rows)}
-        >
-          <BasicGrids />
-          <BookingsLayer />
+          {/* Calendar data */}
+          <div
+            data-role='calendar-data-container'
+            className='relative'
+            style={styleGenerator(CELL_WIDTH_PX * 8, CELL_HEIGHT_PX * rows)}
+          >
+            <BasicGrids />
+            <BookingsLayer containerRef={containerRef} />
+          </div>
         </div>
       </div>
     </div>
-  </div>
-);
+  );
+};
 
 export default Main;
